app_haiiassessment_v2/requirements.txt
Flask==2.1.2
matplotlib
numpy
panda
scipy
seaborn==0.12.2
Werkzeug==2.3.4
